tratando erro se o id não existir

+++ forum/src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -16,6 +16,7 @@ import javax.transaction.Transactional;
 import javax.validation.Valid;
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping(value = "/topicos")
@@ -47,22 +48,33 @@ public class TopicosController {
     }
 
     @GetMapping("/{id}")
-    public DetalheTopicoDto detalhar(@PathVariable Long id){
-       Topico topico = topicoRepository.getReferenceById(id);
-       return new DetalheTopicoDto(topico);
+    public ResponseEntity<DetalheTopicoDto> detalhar(@PathVariable Long id){
+       Optional<Topico> topico = topicoRepository.findById(id);
+       if(topico.isPresent()){
+           return  ResponseEntity.ok(new  DetalheTopicoDto(topico.get()));
+       }
+        return ResponseEntity.notFound().build();
     }
 
     @PutMapping("/{id}")
     @Transactional
     public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizarTopicoForm atualizarTopicoForm) {
-        Topico topico = atualizarTopicoForm.atualizar(id, topicoRepository);
-        return ResponseEntity.ok(new TopicoDto(topico));
+        Optional<Topico> optional = topicoRepository.findById(id);
+        if(optional.isPresent()){
+            Topico topico = atualizarTopicoForm.atualizar(id, topicoRepository);
+            return ResponseEntity.ok(new TopicoDto(topico));
+        }
+        return ResponseEntity.notFound().build();
     }
 
     @DeleteMapping("/{id}")
     @Transactional
     public  ResponseEntity remover(@PathVariable Long id) {
-        topicoRepository.deleteById(id);
-        return ResponseEntity.ok().build();
+        Optional<Topico> topico = topicoRepository.findById(id);
+        if(topico.isPresent()){
+            topicoRepository.deleteById(id);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 }
